+++ src/config.ts
@@ -159,12 +159,12 @@ export const Config: ConfigType = {
         "mastodon.online",
         "mas.to",
         "mastodon.world",
+        "loforo.com",
         "c.im",
         "hachyderm.io",
         "infosec.exchange",
         "universeodon.com",
         "kolektiva.social",
-        "loforo.com",
         "mastodon.gamedev.place",
         "mastodonapp.uk",
         "ioc.exchange",
@@ -180,6 +180,8 @@ export const Config: ConfigType = {
         "metalhead.club",
         "sciences.social",
         "toot.io",
+        "mastodon.ie",
+        "mastodon.nz",
         // Servers that are no bueno for various reasons
         // "baraag.net",                 // very NSFW (anime porn)
         // "mstdn.social",               // Slow, blocked by CORS
@@ -197,17 +199,25 @@ export const Config: ConfigType = {
             "ruhr.social",
             "muenchen.social",
             "social.cologne",
+            "social.tchncs.de",
+            "sueden.social",
+            "mastodontech.de",
+            "nerdculture.de",
         ],
         "es": [
             "tkz.one",
             "mast.lat",
             "mastorol.es",
         ],
+        "eu": [  // Basque language
+            "mastodon.eus",
+        ],
         "fr": [
             "piaille.fr",
             "pouet.chapril.org",
             "mastoot.fr",
             "mamot.fr",
+            "qlub.social", // Montreal
         ],
         "ja": [
             "mstdn.jp",
@@ -227,10 +237,16 @@ export const Config: ConfigType = {
         "ru": [
             "pravda.me",
         ],
+        "tr": [
+            "mastoturk.org",
+        ],
         "zh-cn": [
             "m.cmx.im",
             "m.otter.homes",
             "mast.dragon-fly.club",
+            "alive.bar",
+            "g0v.social",
+            "link.baai.ac.cn",
         ],
     },
     // Non-mastodon servers and/or servers that don't make the MAU data available publicly